Driver app for an implementation of a generic fixed dict.
Fixed Dictionary implementation in C as well as exercises from "Programming in C" by Stephen Kochan (p. 117). Client code in Source.c serves as a test. Example session:

>> nine one one
9 1 1
>> three six niner
3 6
We apologize, but niner was not found in the dictionary.
>> three six nine
3 6 9
>> pistachio cheesecake
We apologize, but pistachio was not found in the dictionary.

We apologize, but cheesecake was not found in the dictionary.